reformatted unit test folder tree
